# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cluvi API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  /** @nullable */
  displayName?: string | null;
  isPro: boolean;
  xp: number;
  level: number;
  streak: number;
  /** @nullable */
  lastActiveDate?: string | null;
  totalPacks: number;
  totalFlashcardsStudied: number;
  totalQuizzesTaken: number;
  createdAt: string;
}

export interface UserProfileUpdate {
  displayName?: string;
}

export interface UserStats {
  xp: number;
  level: number;
  streak: number;
  /** 0-100 study strength score */
  studyStrength: number;
  /** 0-100 exam readiness score */
  examReadiness: number;
  totalPacks: number;
  totalFlashcardsStudied: number;
  totalQuizzesTaken: number;
  averageQuizScore: number;
  weeklyXP: number;
  xpToNextLevel: number;
}

export type StudyPackSourceType =
  (typeof StudyPackSourceType)[keyof typeof StudyPackSourceType];

export const StudyPackSourceType = {
  text: "text",
  pdf: "pdf",
  image: "image",
} as const;

export type StudyPackStatus =
  (typeof StudyPackStatus)[keyof typeof StudyPackStatus];

export const StudyPackStatus = {
  processing: "processing",
  ready: "ready",
  error: "error",
} as const;

export interface StudyPack {
  id: number;
  userId: number;
  title: string;
  sourceType: StudyPackSourceType;
  status: StudyPackStatus;
  /** @nullable */
  summary?: string | null;
  flashcardCount: number;
  quizCount: number;
  examPredictionCount: number;
  /** 0-100 mastery score for this pack */
  studyStrength: number;
  /** Whether this pack has pro features unlocked */
  isPro: boolean;
  createdAt: string;
  updatedAt: string;
}

export type StudyPackInputSourceType =
  (typeof StudyPackInputSourceType)[keyof typeof StudyPackInputSourceType];

export const StudyPackInputSourceType = {
  text: "text",
  pdf: "pdf",
  image: "image",
} as const;

export interface StudyPackInput {
  /** @minLength 1 */
  title: string;
  sourceType: StudyPackInputSourceType;
  /** Raw text content or base64-encoded file */
  content: string;
}

export type StudyPackFullSourceType =
  (typeof StudyPackFullSourceType)[keyof typeof StudyPackFullSourceType];

export const StudyPackFullSourceType = {
  text: "text",
  pdf: "pdf",
  image: "image",
} as const;

export type StudyPackFullStatus =
  (typeof StudyPackFullStatus)[keyof typeof StudyPackFullStatus];

export const StudyPackFullStatus = {
  processing: "processing",
  ready: "ready",
  error: "error",
} as const;

export interface Flashcard {
  id: number;
  studyPackId: number;
  front: string;
  back: string;
  /** 1-5 confidence rating from user */
  confidenceLevel: number;
  /** @nullable */
  nextReviewAt?: string | null;
  timesStudied: number;
  createdAt: string;
}

export interface QuizQuestion {
  id: number;
  studyPackId: number;
  question: string;
  options: string[];
  /** Index of correct option (0-based) */
  correctAnswer: number;
  /** @nullable */
  explanation?: string | null;
  createdAt: string;
}

export interface StudyPackFull {
  id: number;
  userId: number;
  title: string;
  sourceType: StudyPackFullSourceType;
  status: StudyPackFullStatus;
  /** @nullable */
  summary?: string | null;
  keyConcepts?: string[];
  examPredictions?: string[];
  flashcardCount: number;
  quizCount: number;
  examPredictionCount: number;
  studyStrength: number;
  isPro: boolean;
  flashcards: Flashcard[];
  quizQuestions: QuizQuestion[];
  createdAt: string;
  updatedAt: string;
}

export interface DashboardSummary {
  totalPacks: number;
  recentPacks: StudyPack[];
  xp: number;
  level: number;
  streak: number;
  studyStrength: number;
  examReadiness: number;
  averageQuizScore: number;
  weakTopics: string[];
  totalFlashcardsStudied: number;
  totalQuizzesTaken: number;
}

export interface ConfidenceRating {
  /**
   * @minimum 1
   * @maximum 5
   */
  confidence: number;
}

export type QuizSubmissionAnswersItem = {
  questionId: number;
  selectedAnswer: number;
};

export interface QuizSubmission {
  answers: QuizSubmissionAnswersItem[];
  timeTakenSeconds?: number;
}

export interface QuizResult {
  id: number;
  studyPackId: number;
  /** Percentage score 0-100 */
  score: number;
  correctCount: number;
  totalQuestions: number;
  /** @nullable */
  timeTakenSeconds?: number | null;
  xpEarned: number;
  weakAreas: string[];
  recommendations: string[];
  createdAt: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface GeminiConversationInput {
  title: string;
}

export interface GeminiMessageInput {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GeminiImageInput {
  prompt: string;
}

export interface GeminiImageOutput {
  b64_json: string;
  mimeType: string;
}

export interface GeminiError {
  error: string;
}